requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
fastapi==0.63.0
h11==0.12.0
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.3
MarkupSafe==1.1.1
motor==2.3.0
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.1
pymongo==3.11.2
python-multipart==0.0.5
pytz==2020.5
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
Werkzeug==1.0.1
win32-setctime==1.0.3
wincertstore==0.2
docx2txt
